requirements.txt
requests
pyyaml
pandas==0.24.1